pyproject.toml
[project]
name = "ac"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "anthropic>=0.86.0",
    "groq>=1.0.0",
    "ipykernel>=6.31.0",
    "python-dotenv>=1.2.1",
    "rich>=14.3.3",
]
